Flush out more interfaces
