+++ Kava/Kava/Source/PageObjects/Infrastructure/Elements/CollectionView.swift
@@ -1,0 +1,20 @@
+//
+//  CollectionView.swift
+//  Kava
+//
+//  Created by John Hammerlund on 3/8/16.
+//  Copyright © 2016 MINDBODY, Inc. All rights reserved.
+//
+
+import Foundation
+import XCTest
+
+public class CollectionView<TResultBlock: Block> : Swipeable<TResultBlock> {
+    
+    public required init(parentBlock: Block, backingElement: XCUIElement) {
+        super.init(parentBlock: parentBlock, backingElement: backingElement)
+    }
+    
+    
+    
+}
